Preprocess data
Began preprocessing the data, choosing a suitable subset for use in the classification problem. Only the high quality recordings in the Texas area will be used, and the number of classes to identify was decreased so that all classes had at least 5 minutes worth of audio data. Further processing of the audio data is required.