Remove signal-free sections
I filtered out the silence/noise sections of the audio data by comparing the largest signals to the base noise level. Noise during the signals is still not filtered out, but this should significantly help with identification, since the model doesn't have to try to identify something in a section of silence or other background birds.