Try models on raw data
This was suggested to me after I found that normalizing was lowering accuracy, but the raw data based models were just as bad. I was kind of expecting this, but I'm curious now whether I can teach a model to do Fourier transforms. At this point the next step is to either try to apply noise removal and/or add more of the lower quality data in.